♻️ Extract the data models into a single module to manage and maintain.
